Run clean ups with removeOrphanedEvents in major batches
It's already wrapped in a small batch but if the table is large, this
would not work.

Adding a major batch to limit the rows scanned drastically makes it
faster. I have been running this in production and it works like a charm
on a table with > 300M rows

Bug: T310428
Change-Id: I7f63d2d76fea0db64e48dbcc1f95419df4492e7d
(cherry picked from commit 6d0e181bf4224a94164b6966429febf4578d83b2)